requirements.txt
pandas
numpy
mplfinance
psycopg2-binary
python-dotenv